paypal/requirements.txt
paypal-agent-toolkit==1.3.0
# CrewAI
crewai==0.11.0
crewai-tools==0.13.2
langchain==0.1.20
langchain-openai==0.0.2.post1
python-dotenv>=1.0.1
openai>=1.0.0
